requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
python-dotenv==1.0.0

Werkzeug~=3.1.3
WTForms~=3.2.1